Add a stDebugProcessRunning node to allow editor to not be readonly
The plugin noderunner defines <a:state id="stDebugProcessRunning"> which is
references by code.xml when deciding whether to set the editor to readonly. If
this is not present, code.xml seems to default to readonly. By including it
without any functionality, the editor is free to work properly. Oh APF, you
never cease to bring me joy.